Move Citation to shared API model types.
