+++ src/shared/types/apiModel.ts
@@ -15,3 +15,19 @@ export type Xref = {
   isoformId?: string;
   implicit?: true;
 };
+
+export type Citation = {
+  citationType?: string;
+  authors?: string[];
+  citationCrossReferences?: Xref[];
+  title?: string;
+  publicationDate?: string;
+  journal?: string;
+  firstPage?: string;
+  lastPage?: string;
+  volume?: string;
+  completeAuthorList?: boolean;
+  literatureAbstract?: string;
+  authoringGroup?: string[];
+  submissionDatabase?: string;
+};
